Updated Models and fixed path errors

+++ store_closure/models.py
@@ -26,15 +26,15 @@ class QueryStatus(models.Model):
 
 
 class Homedata(models.Model):
-    longitude = models.DecimalField(max_digits = 18 , decimal_places=15)
-    latitude = models.DecimalField(max_digits = 18, decimal_places=15)
+    longitude = models.FloatField()
+    latitude = models.FloatField()
     hhinc8 = models.IntegerField()
     owncar = models.IntegerField()
     category = models.CharField(max_length=4)
     salary = models.IntegerField()
 
 class Marketdata(models.Model):
-    longitude = models.DecimalField(max_digits = 18 , decimal_places=15)
-    latitude = models.DecimalField(max_digits = 18, decimal_places=15)
+    longitude = models.FloatField()
+    latitude = models.FloatField()
     category = models.CharField(max_length=4)
     FSA = models.IntegerField()
